DS-4668 update main from develop for 4.3.0 release (#45)
* [DS-4441] MacOS Setup for docker-compose swirl env (#40)

* chore(infra): add variable export to find docker binary

* chore(infra): point log directory to home due to launchctl permissions

* feat(infra): implements logic for running swirl as a service in macos devices

* chore(infra): resolve docker commands to use binary variable

* chore(infra): specify docker compose file for dynamically starting service

* feat(infra): include documentation for swirl macos service

* fix(infra): remove duplicated patch for linux swirl service file

* chore(infra): update replaced snippets to match existing pattern

* chore(infra): include dynamic creation for base log directory in macos and ubuntu

---------


* [DS-3709] Update docker compose to reload prompts always (#41)

* chore(infra): update swirl_app_init command to reload ai prompts by default

* chore(infra): update swirl-init to use swirl.py instead of manage.py

* feat(infra): include swirl-reload service for reloading ai prompts automatically

* chore(infra): include reload service to start as a profile

* chore(infra): include reload ai prompts command to swirl load job

---------


* [DS-4030] Add Health Check Script to docker image (#42)

* chore(infra): remove swirl health check script

* chore(infra): update swirl health check to use script inside /app folder

---------


* handle stop more completely

* - update example for 4.3
- fix lack f proper cleanup in dev tool
- add branding to job
- fix flaw in example copy

---------

+++ DevUtils/reset_compose_environment.sh
@@ -2,14 +2,17 @@
 
 
 # Stop Swirl
-systemctl stop swirl
+if [[ "$OSTYPE" == "linux-gnu"* ]]; then
+  systemctl stop swirl
+fi
 
+docker-compose --profile all down
 # Remove Swirl's Docker containers
 docker system prune --force
 
 # Remove Volumes
 docker volume prune --force
-docker volume rm app_db_data
+docker volume rm swirl_db_data
 
 # Verify that the volumes and containers are removed
 docker volume ls
